requirements.txt
pandas
pyarrow 
torch
ib_insync 
numpy
pyyaml
flask
openpyxl
numba
tqdm
matplotlib
pypdf2
cuda12x
scikit-learn
lightgbm